Feat: Extract ModalHeaderContents into a separate component
