+++ src/components/admin/order/OrderDetailModal.tsx
@@ -1,11 +1,10 @@
-import { Order, OrderStatus } from '@@types/index';
+import { Order } from '@@types/index';
 import RoundedAppButton from '@components/common/button/RoundedAppButton';
 import AppLabel from '@components/common/label/AppLabel';
 import styled from '@emotion/styled';
 import { Color } from '@resources/colors';
-import CloseSvg from '@resources/svg/CloseSvg';
-import { expandButtonStyle } from '@styles/buttonStyles';
 import { colFlex, rowFlex } from '@styles/flexStyles';
+import ModalHeaderContents from './ModalHeaderContents';
 
 const ModalOverlay = styled.div`
   position: fixed;
@@ -36,28 +35,6 @@ const ModalContainer = styled.div`
 const HorizontalLine = styled.hr`
   width: 100%;
   border: 0.3px solid #eeecec;
-`;
-
-const ModalHeader = styled.div`
-  ${colFlex({ justify: 'space-between', align: 'start' })}
-  gap: 10px;
-  padding-bottom: 10px;
-  border-bottom: 20px solid ${Color.LIGHT_GREY};
-`;
-
-const ModalHeaderTitle = styled.div`
-  ${rowFlex({ justify: 'space-between', align: 'center' })}
-  width: 100%;
-`;
-
-const HeaderDetailContainer = styled.div`
-  ${rowFlex({ justify: 'space-between', align: 'center' })}
-  width: 100%;
-`;
-
-const HeaderDetail = styled.div`
-  ${colFlex({ align: 'start' })}
-  gap: 10px;
 `;
 
 const ModalContent = styled.div`
@@ -95,22 +72,11 @@ const ModalFooter = styled.div`
   width: 100%;
 `;
 
-const CloseIcon = styled(CloseSvg)`
-  ${expandButtonStyle}
-`;
-
 interface OrderDetailModalProps {
   isOpen: boolean;
   orderInfo: Order;
   onClose: () => void;
 }
-
-const orderStatusMap: Record<OrderStatus, string> = {
-  [OrderStatus.NOT_PAID]: '주문 완료',
-  [OrderStatus.PAID]: '결제 완료',
-  [OrderStatus.SERVED]: '서빙 완료',
-  [OrderStatus.CANCELLED]: '취소됨',
-};
 
 function OrderDetailModal({ isOpen, onClose, orderInfo }: OrderDetailModalProps) {
   if (!isOpen) return null;
@@ -119,21 +85,7 @@ function OrderDetailModal({ isOpen, onClose, orderInfo }: OrderDetailModalProps)
     <>
       <ModalOverlay onClick={onClose} />
       <ModalContainer>
-        <ModalHeader>
-          <ModalHeaderTitle>
-            <AppLabel color={Color.BLACK} size={17} style={{ fontWeight: 800 }}>{`테이블 ${orderInfo.tableNumber + 1}`}</AppLabel>
-            <CloseIcon onClick={onClose} />
-          </ModalHeaderTitle>
-          <HeaderDetailContainer>
-            <HeaderDetail>
-              <AppLabel color={Color.BLACK} size={13}>{`주문 번호  ${orderInfo.id + 1}`}</AppLabel>
-              <AppLabel color={Color.BLACK} size={13}>{`주문 일시 | ${new Date(orderInfo.createdAt).toLocaleString()}`}</AppLabel>
-            </HeaderDetail>
-            <AppLabel color={Color.KIO_ORANGE} size={17} style={{ fontWeight: 600 }}>
-              {orderStatusMap[orderInfo.status]}
-            </AppLabel>
-          </HeaderDetailContainer>
-        </ModalHeader>
+        <ModalHeaderContents onClose={onClose} orderInfo={orderInfo} />
         <ModalContent>
           <AppLabel color={Color.BLACK} size={20} style={{ fontWeight: 700 }}>
             주문 내역
